+++ ginger.py
@@ -14,12 +14,19 @@ Options:
   -h, --help          Show this screen.
 
 Input formats:
-  - `guess`      Try to guess the file format, defaults to CoNLL-U
-  - `conllx`     [CoNLL-X format](https://web.archive.org/web/20160814191537/http://ilk.uvt.nl:80/conll/)
-  - `conllu`     [CoNLL-U format](http://universaldependencies.org/format.html)
-  - `conll2009`  [CoNLL-2009 format](http://ufal.mff.cuni.cz/conll2009-st/task-description.html)
-  - `talismane`  Outputs of [Talismane](http://redac.univ-tlse2.fr/applications/talismane/talismane_en.html)
-  - `mate`       Alias for `conll2009`, used by [mate-tools](http://www.ims.uni-stuttgart.de/forschung/ressourcen/werkzeuge/matetools.en.html)
+  - `guess`           Try to guess the file format, defaults to CoNLL-U
+  - `conllx`          [CoNLL-X format](https://web.archive.org/web/20160814191537/http://ilk.uvt.nl:80/conll/)
+  - `conllu`          [CoNLL-U format](http://universaldependencies.org/format.html)
+  - `conll2009_gold`  [CoNLL-2009 format](http://ufal.mff.cuni.cz/conll2009-st/task-description.html)
+                      (Gold columns only)
+  - `conll2009_sys`   [CoNLL-2009 format](http://ufal.mff.cuni.cz/conll2009-st/task-description.html)
+                      (Predicted columns only)
+  - `talismane`       Outputs of [Talismane](http://redac.univ-tlse2.fr/applications/talismane/talismane_en.html)
+  - `mate_gold`       Alias for `conll2009_gold`, used by
+                      [mate-tools](http://www.ims.uni-stuttgart.de/forschung/ressourcen/werkzeuge/matetools.en.html)
+  - `mate_sys`        Alias for `conll2009_sys`, used by
+                      [mate-tools](http://www.ims.uni-stuttgart.de/forschung/ressourcen/werkzeuge/matetools.en.html)
+
 
 Output formats
   - Text formats. Those can't be used without any dependency
@@ -31,16 +38,15 @@ Example:
   `ginger -f conllu input.conll -t tikz output.tex`
 """
 
-__version__ = 'ginger 0.5.0'
+__version__ = 'ginger 0.7.0'
 
 import sys
 import contextlib
 from docopt import docopt
 
-import re
-
 import signal
 signal.signal(signal.SIGPIPE, signal.SIG_DFL)
+
 
 import logging
 logging.basicConfig(level=logging.INFO)
@@ -53,6 +59,14 @@ except ImportError:
     from . import libginger
     from . import libtreebank
     from . import libtreerender
+
+
+def sigint_handler(signal, frame):
+        logging.error('Process interrupted by SIGINT')
+        sys.exit(0)
+
+
+signal.signal(signal.SIGINT, sigint_handler)
 
 
 # Thanks http://stackoverflow.com/a/17603000/760767
@@ -79,11 +93,11 @@ def main_entry_point(argv=sys.argv[1:]):
     if arguments['<out-file>'] is None:
         arguments['<out-file>'] = '-'
 
-    with smart_open(arguments['<in-file>']) as in_stream:
-        in_str = in_stream.read()
+    with smart_open(arguments['<in-file>'], encoding='utf8') as in_stream:
+        in_lst = list(in_stream.readlines())
 
     if arguments['--from'] == 'guess' or arguments['--from'] is None:
-        arguments['--from'] = libtreebank.guess(in_str)
+        arguments['--from'] = libtreebank.guess(in_lst)
 
     try:
         parser, _ = libtreebank.formats[arguments['--from']]
@@ -97,21 +111,26 @@ def main_entry_point(argv=sys.argv[1:]):
             argsfrom=arguments['--from']))
         sys.exit(1)
 
-    treebank = [parser(tree) for tree in re.split('\n\n+', in_str.strip()) if tree and not tree.isspace()]
+    treebank = parser(in_lst)
 
-    # Binary outputs
-    if arguments['--to'] in ['png']:
+    # Cairo-based outputs
+    if arguments['--to'] in ('png', 'svg'):
         if arguments['--to'] == 'png':
             out_bytes = libtreerender.to_png(treebank[0])
+        if arguments['--to'] == 'svg':
+            out_bytes = libtreerender.to_svg(treebank[0])
         with open(arguments['<out-file>'], 'wb') as out_stream:
             out_stream.write(out_bytes)
-            # Text outputs
+    # Text outputs
     else:
+        # Text-based graphics
         if arguments['--to'] == 'tikz':
-            out_str = '\n\n'.join(libtreerender.tikz(t) for t in treebank)
+            out_lst = [libtreerender.tikz(t) for t in treebank]
 
         elif arguments['--to'] == 'ascii':
-            out_str = '\n\n'.join(libtreerender.ascii_art(t) for t in treebank)
+            out_lst = [libtreerender.ascii_art(t) for t in treebank]
+
+        # Treebank
         else:
             try:
                 _, formatter = libtreebank.formats[arguments['--to']]
@@ -125,11 +144,14 @@ def main_entry_point(argv=sys.argv[1:]):
                 argsto=arguments['--to']))
                 sys.exit(1)
 
-            out_str = '\n\n'.join(formatter(t) for t in treebank)
+            out_lst = [formatter(t) for t in treebank]
 
-        with smart_open(arguments['<out-file>'], 'w') as out_stream:
-            out_stream.write(out_str)
-            out_stream.write('\n')
+    with smart_open(arguments['<out-file>'], 'w', encoding='utf8') as out_stream:
+        for t in out_lst[:-1]:
+            out_stream.write(t)
+            out_stream.write('\n\n')
+        out_stream.write(out_lst[-1])
+        out_stream.write('\n')
 
 
 if __name__ == '__main__':