+++ libginger.py
@@ -5,10 +5,6 @@ import typing as ty
 import itertools as it
 
 
-class ParsingError(Exception):
-    pass
-
-
 class Node:
     '''A node in an UD graph.'''
     def __init__(self,
@@ -109,72 +105,3 @@ class Tree:
 
     def __str__(self):
         return '\n'.join(str(n) for n in self.nodes)
-
-    # TODO: Add a way to deal with the caveats
-    @classmethod
-    def from_conll(cls, conll_str: str):
-        """Return a tree (that is a list of `Nodes`) from its
-           [CoNNL-U string representation](http://universaldependencies.org/format.html).
-
-           ## Caveats
-             - Empty nodes are skipped"""
-        root = Node(identifier=0, form='ROOT')
-        res = [root]
-        # First get the self-contained values, deal with references later
-        # IMPLEMENTATION: This relies on references being initialisable
-        #                 with identifiers
-        for i, line in enumerate(l.strip() for l in conll_str.splitlines()):
-            # Skip comment lines
-            if line.startswith('#'):
-                next
-
-            try:
-                identifier, form, lemma, upostag, xpostag, feats, head, deprel, deps, misc = line.split('\t')
-            except ValueError:
-                # TODO: Issue a warning here
-                raise ParsingError('At line {i} : 10 columns expected, got a {n} ({line!r})'.format(i=i, n=len(line.split('\t')), line=line))
-            try:
-                identifier = int(identifier)
-            except ValueError:
-                # TODO: Issue a warning here
-                if re.match(r'\d+-\d+', identifier):  # Skip multiword tokens
-                    next
-                if re.match(r'\d+.\d+', identifier):  # Skip empty nodes
-                    next
-                raise ParsingError('At line {i} : the `id` field does not respect CoNLL-U specifications.'.format(i=i))
-
-            try:
-                feats = dict() if feats == '_' else dict(e.split('=') for e in feats.split('|'))
-            except ValueError:
-                # Be nice : if empty, it should be an underscore, but let's be nice with spaces and empty strings, too
-                if feats.isspace() or not feats:
-                    # TODO: Issue a warning here
-                    feats = dict()
-                else:
-                    raise ParsingError('At line {i} : the `feats` field does not respect CoNLL-U specifications.'.format(i=i))
-
-            try:
-                head = int(head)
-            except ValueError:
-                raise ParsingError('At line {i} : the `head` field does not respect CoNLL-U specifications.'.format(i=i))
-
-            try:
-                deps = [] if deps == '_' else [e.split(':') for e in deps.split('|')]
-            except ValueError as e:
-                # Be nice : if empty, it should be an underscore, but let's be nice with spaces and empty strings, too
-                if deps.isspace() or not deps:
-                    # TODO: Add a warning here
-                    deps = []
-                else:
-                    raise ParsingError('At line {i} : the `deps` field does not respect CoNLL-U specifications.'.format(i=i))
-
-            new_node = Node(identifier, form, lemma, upostag, xpostag, feats, head, deprel, deps, misc)
-            res.append(new_node)
-
-        # Now deal with references, which is easy, since the index of a node in
-        # `res` is exactly its identifier
-        for n in res[1:]:
-            n.head = res[n.head]
-            n.deps = [(res[head], dep) for head, dep in n.deps]
-
-        return cls(res)